Apps/sprite_sheet_splitter/requirements.txt
Pillow
PySide6
